feat(cipher): CipherState as interface
While working on integrating tests from `snow`, I eventually realized I did a oopsie since the protocol name which decides the noise protocol to use is not known at comptime.

This commit changes `CipherState` to use a tagged union instead. We have a very limited number of ciphers to choose from, so a tagged union seems fair in this case, and simplifies our interface.