feat: specialize `mixHash` into alloc and noalloc versions
We split `mixHash` into alloc and noalloc versions, the alloc version being the original using `std.mem.concat` and the noalloc version uses a `BoundedArray` instead.

We do not always need to rely on `std.mem.concat` for all cases of mixing - in
some cases we mix data with known sizes like with cipher keys (32 bytes) or HKDF output (64 bytes). In this case, we use `mixHashBounded` with a `BoundedArray` of capacity 128.

`mixHash` stays as before, relying on `std.mem.concat` and this would be for either mixing the prologue, or mixing during encryption/decryption, since the payloads for those workloads are variable-length.